requirements.txt
pymysql>=1.1.0
requests>=2.32.0
python-dotenv>=1.0.0
